add compliance calculation worksheet
the user is responsible for confirming that their stimulation levels are within compliance, as this is not included in the checks in the code. the impedances must be measured separately (i.e. using clinical software such as SoundWave), and the compliance limits for each electrode can be calculated using this spreadsheet.